bingo-50/pyproject.toml
[tool.poetry]
name = "bingo-50"
version = "0.1.0"
description = ""
authors = ["PTMGitHub <61422343+PTMGitHub@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
